==> 4.relation_extraction.ipynb <==
import pandas as pd
import transformers

import torch

device=torch.device("cpu")
#device=torch.device("cuda")

from transformers import BertTokenizer
from transformers import BertForSequenceClassification
tokenizer = BertTokenizer.from_pretrained('Repository/biobert/tokenizer_biobert', do_lower_case=True)
model = BertForSequenceClassification.from_pretrained(
     "Repository/biobert/model_biobert-finetuning",
     num_labels = 2, 
     output_attentions = False, # Whether the model returns attentions weights.
     output_hidden_states = True, # Whether the model returns all hidden-states.
).to(device)

import tqdm
model.eval()
with torch.no_grad():
 f = open("data/masked_sentences.txt","r")
 f2 = open("data/masked_sentences_bert.txt","w")
 m = torch.nn.Softmax()
 for line in tqdm.tqdm(f):
   ls = line.strip().split("\t")
   sent = ls[5]
   if len(sent.split(" ")) > 500:continue
   indexed_tokens = tokenizer.encode(
                        sent,                     # Sentence to encode.
                        add_special_tokens = True)
   tokens_tensor = torch.tensor([indexed_tokens])
   if len(tokens_tensor[0]) > 500:continue
   all_encoder_layers = model(tokens_tensor.to(device))
   out= m(all_encoder_layers[0])
   judge = int(out[0][1] > 0.5)
   if judge ==1:
     f2.write(str(out[0][1].item())+"\t"+line)
 f2.close()
