==> 6.feature.ipynb <==
import pandas as pd
import numpy as np
import gene_to_sym
import tqdm
gene_to_sym = gene_to_sym.get()

#chromosome
sym,chr = [],[]
f = open("data/HGNC_0908.txt","r")
for line in f:
  if line.startswith("Approved"):continue
  ls = line.strip("\n").split("\t")
  sym.append(str(ls[0]))
  chr.append(str(ls[4]))
f.close()

chrdic = {}
for symbol in gene_to_sym.values():
    chrdic[symbol] = set()
    
chrset = set()
for n,s in enumerate(sym):
  s = s.upper()
  if "p" in chr[n]:
     c = chr[n][:chr[n].index("p")+1]
  elif "q" in chr[n]:
     c = chr[n][:chr[n].index("q")+1]
  else:continue
  if len(c) > 4:continue
  if c == "unp":continue
  chrset.add(c)
  chrdic[s]=set([c])
chrlist = sorted(list(chrset))
chrnum = {}
for n,e in enumerate(chrlist):
  chrnum[e] = n

#!wget http://geneontology.org/gene-associations/goa_human.gaf.gz
#!gunzip goa_human.gaf.gz
dfgo = pd.read_table("Repository/Feature/goa_human.gaf",skiprows=41,header=None)
godic = {}
for symbol in gene_to_sym.values():
    godic[symbol] = set()
name = dfgo[2].tolist()
goid = dfgo[4].tolist()
goset= set(goid)
for n in range(len(dfgo)):
  gene = str(name[n]).upper()
  if gene in gene_to_sym:
    symbol = gene_to_sym[gene]
    godic[symbol].add(goid[n])
golist = sorted(list(goset))
gonum = {}
for n in range(len(golist)):
   gonum[golist[n]]=n

#MGI
#!wget http://www.informatics.jax.org/downloads/reports/HMD_HumanPhenotype.rpt
f = open("Repository/Feature/HMD_HumanPhenotype.rpt","r")
mgidic = {}
mgiset = set()
for symbol in gene_to_sym.values():
    mgidic[symbol] = set()
for line in f:
  ls = line.strip().split("\t")
  if len(ls) >= 5:
      gene = ls[0].upper()
      if gene not in gene_to_sym:
        continue
      mgidic[gene_to_sym[gene]]= set(ls[4].strip().split(", "))
      mgiset = mgiset | mgidic[gene_to_sym[gene]]
mgilist = list(mgiset)
mgilist.sort()
mginum = {}
for t in range(len(mgilist)):
    mginum[mgilist[t]] =t
f.close()

#HPO,OMIM,ORPHAN
#!wget http://purl.obolibrary.org/obo/hp/hpoa/genes_to_phenotype.txt

f = open("Repository/Feature/genes_to_phenotype.txt", "r")
a = 0
hpodic = {}
ormdic = {}
hposet = set()
ormset = set()
for symbol in gene_to_sym.values():
    hpodic[symbol] = set()
    ormdic[symbol] = set()
for line in f:
  if a == 0:
    a += 1
    continue
  ls = line.strip().split("\t")
  gene = ls[1].upper()
  if gene not in gene_to_sym:continue
  symbol = gene_to_sym[gene]
  hpodic[symbol].add(ls[2])
  hposet.add(ls[2])
  ormdic[symbol].add(ls[8])
  ormset.add(ls[8])

p = sorted(list(hposet))
o = sorted(list(ormset))
hponum = {}
ormnum = {}
for t in range(len(p)):
    hponum[p[t]] =t 
for t in range(len(o)):
    ormnum[o[t]] =t 

#Transcript factors
#!wget https://maayanlab.cloud/static/hdfs/harmonizome/data/encodetfppi/gene_attribute_matrix.txt.gz
#!gunzip gene_attribute_matrix.txt.gz

df = pd.read_table('Repository/Feature/gene_attribute_matrix.txt')

dftf = df.loc[2:].copy()
dftf.index=dftf["#"]
dftf=dftf.drop(["#","#.1","GeneSym"],axis=1)
dftf = dftf.astype('float16')
tfdic = {}
for symbol in gene_to_sym.values():
    tfdic[symbol] = set()

    
ind = dftf.index.tolist()
col = dftf.columns.tolist()

for column in tqdm.tqdm(col):
  row = dftf[column].tolist()
  for n in range(len(row)):
    if row[n] == 1.0:
      if ind[n].upper() in gene_to_sym:
         tfdic[gene_to_sym[ind[n].upper()]].add("TF:"+column)
tfnum={}
t = 0
for g in col:
  tfnum["TF:"+g] = t
  t += 1

#Biogrid
#!wget https://downloads.thebiogrid.org/Download/BioGRID/Release-Archive/BIOGRID-4.2.193/BIOGRID-ALL-4.2.193.tab3.zip
#!unzip BIOGRID-ALL-4.2.193.tab3.zip
f = open("Repository/Feature/BIOGRID-ALL-4.2.193.tab3.txt","r")
intdic={}
intnum={}
for symbol in gene_to_sym.values():
    intdic[symbol] = set()
for line in f:
  ls = line.strip().split("\t")
  if ls[-2] == "Homo sapiens" and ls[-1] =="Homo sapiens":
      a = ls[7].upper()
      b = ls[8].upper()
      if a in gene_to_sym and b in gene_to_sym:
        a = gene_to_sym[a]
        b = gene_to_sym[b]
        intdic[a].add("BG:"+b)
        intdic[b].add("BG:"+a)
ints = sorted(["BG:"+i for i in list(intdic.keys())])
for n,ii in enumerate(ints):
  intnum[ii] = n

#subcellular localization
#!wget https://www.proteinatlas.org/download/subcellular_location.tsv.zip
#!unzip subcellular_location.tsv.zip
import pandas as pd
slodic = {}
for symbol in gene_to_sym.values():
    slodic[symbol] = set()
df = pd.read_table('Repository/Feature/subcellular_location.tsv')
location = df["Main location"].tolist()
genes = df["Gene name"].tolist()
sloset = set()
for n in range(len(df)):
    gene = genes[n].upper()
    if gene not in gene_to_sym:
      continue
    symbol = gene_to_sym[gene]
    slodic[symbol] = slodic[symbol] | set(location[n].strip().split(";"))
    sloset = sloset | set(location[n].strip().split(";"))
slolist = sorted(list(sloset))
slonum = {}
for n,slo in enumerate(slolist):
  slonum[slo] = n

#Expression
#!wget https://www.proteinatlas.org/download/rna_tissue_hpa.tsv.zip
#!unzip rna_tissue_hpa.tsv.zip
dfrna = pd.read_table("Repository/Feature/rna_tissue_hpa.tsv")
grouped = dfrna.groupby('Gene name')
meandic = {}
stddic = {}
gm = grouped.mean()["TPM"].tolist()
gi = grouped.mean()["TPM"].index.tolist()
gs = grouped.std()["TPM"].tolist()
for n in range(len(gm)):
  meandic[gi[n]] = gm[n]
  stddic[gi[n]] = gs[n]

tisdic={}
for symbol in gene_to_sym.values():
    tisdic[symbol] = set()

genes = dfrna["Gene name"].tolist()
tpms = dfrna["TPM"].tolist()
relative_expression = []
for n in tqdm.tqdm(range(len(dfrna))):
  gene = genes[n]
  relative_expression.append((tpms[n]-meandic[genes[n]])/(stddic[genes[n]]+0.001))
tissues = dfrna["Tissue"].tolist()

for n in tqdm.tqdm(range(len(dfrna))):
  g = genes[n].upper()
  if g not in gene_to_sym:
    continue
  t = tissues[n]
  rel = relative_expression[n]
  sym = gene_to_sym[g]
  if rel >= 1.68: tisdic[sym].add(t+":H5")
  if  0.91 <= rel <  1.68: tisdic[sym].add(t+":H4")
  if  0.53 <= rel <  0.91: tisdic[sym].add(t+":H3")
  if  0.28 <= rel <  0.53: tisdic[sym].add(t+":H2")
  if  0.09 <= rel <  0.28: tisdic[sym].add(t+":H1")
  if -0.16 <= rel < -0.25: tisdic[sym].add(t+":L1")
  if -0.25 <= rel < -0.39: tisdic[sym].add(t+":L2")
  if -0.39 <= rel < -0.25: tisdic[sym].add(t+":L3")
  if -0.65 <= rel < -0.39: tisdic[sym].add(t+":L4")
  if rel < -0.65: tisdic[sym].add(t+":L5")
tisnum = {}
c = 0
for t in sorted(list(set(tissues))):
  for n in range(1,6):
    tisnum[t+":H"+str(n)]=c
    c +=1
  for n in range(1,6):
    tisnum[t+":L"+str(n)]=c
    c +=1

#Depmap
#!wget https://ndownloader.figshare.com/files/31315996
#%mv 31315996 Crispr.csv

dfdep = pd.read_csv("Repository/Feature/Crispr.csv")
dfdep = dfdep.drop("DepMap_ID",axis=1)
dfdep = dfdep.dropna()
dfdep.columns = [d[:d.index("(")-1].upper() for d in dfdep.columns]
c = [a.upper() in gene_to_sym for a in dfdep.columns]
dfdep = dfdep.loc[:,c]
dfdep.columns = [gene_to_sym[a.upper()] for a in dfdep.columns]
depmap_dict = {}
for d in tqdm.tqdm(dfdep.columns):
  depmap_dict[d] = np.array(dfdep[d])

#word2vec
import pickle
with open("Repository/Feature/gene2vec.pickle","rb") as f:
  gene2vec = pickle.load(f)

alldic = {}
for symbol in gene_to_sym.values():
    alldic[symbol] = set()

def makedic(d):
  for symbol in gene_to_sym.values():
    alldic[symbol] = alldic[symbol] | d[symbol]
  return alldic
alldic = makedic(godic)
alldic = makedic(mgidic)
alldic = makedic(slodic)
alldic = makedic(hpodic)
alldic = makedic(ormdic)
alldic = makedic(intdic)
alldic = makedic(tfdic)
alldic = makedic(tisdic)
alldic = makedic(chrdic)

allnum = {}
def makenum(dn):
  a = len(allnum)
  for k in dn:
     allnum[k] = a+ dn[k]
  return allnum

allnum = makenum(gonum)
allnum = makenum(mginum)
allnum = makenum(slonum)
allnum = makenum(hponum)
allnum = makenum(ormnum)
allnum = makenum(intnum)
allnum = makenum(tfnum)
allnum = makenum(tisnum)
allnum = makenum(chrnum)

import pickle
with open("model/alldic.pickle", "wb") as f:
  pickle.dump(alldic,f)
with open("model/allnum.pickle", "wb") as f:
  pickle.dump(allnum,f)

with open("model/depmap.pickle", "wb") as f:
  pickle.dump(depmap_dict,f)

datdic = {}
for symbol in gene_to_sym.values():
    datdic[symbol] = set()

def makedic(d):
  for symbol in gene_to_sym.values():
    datdic[symbol] = datdic[symbol] | d[symbol]
  return datdic

datdic = makedic(mgidic)
datdic = makedic(slodic)
datdic = makedic(hpodic)
datdic = makedic(intdic)
datdic = makedic(tfdic)
datdic = makedic(tisdic)
datdic = makedic(chrdic)

datnum = {}
def makenum(dn):
  a = len(datnum)
  for k in dn:
      datnum[k] = a+ dn[k]
  return datnum

datnum = makenum(mginum)
datnum = makenum(slonum)
datnum = makenum(hponum)
datnum = makenum(intnum)
datnum = makenum(tfnum)
datnum = makenum(tisnum)
datnum = makenum(chrnum)

import pickle
with open("model/datdic.pickle", "wb") as f:
  pickle.dump(datdic,f)
with open("model/datnum.pickle", "wb") as f:
  pickle.dump(datnum,f)
